+++ CHANGELOG.md
@@ -5,8 +5,12 @@
 #### Bugs
 * Fix #2271: Support periodic refresh of access tokens before they expire
 * Fix #3733: The authentication command from the .kube/config won't be discarded if no arguments are specified
+* * Fix #4206: KubernetesDeserializer can now handle any valid object. If the object lacks type information, it will be deserialized as a GenericKubernetesResource.
 * Fix #4365: backport of stopped future for informers to obtain the termination exception
 * Fix #4383: bump snakeyaml from 1.28 to 1.33
+
+#### _**Note**_: Behavior changes
+* Fix #4206: The Serialization utility class will throw an Exception, instead of returning null, if an untyped unmarshall method is used on something that lacks type information
 
 ### 5.12.3 (2022-07-27)